+++ tests/test_launcher.py
@@ -1,6 +1,34 @@
-import pytest
+import re
 
-from scrapyd.launcher import get_crawl_args
+import pytest
+from twisted.internet import error
+from twisted.logger import LogLevel, capturedLogs, eventAsText
+from twisted.python import failure
+
+from scrapyd import __version__
+from scrapyd.config import Config
+from scrapyd.interfaces import IEnvironment
+from scrapyd.launcher import Launcher, get_crawl_args
+
+
+def message(captured):
+    return eventAsText(captured[0]).split(" ", 1)[1]
+
+
+@pytest.fixture()
+def environ(app):
+    return app.getComponent(IEnvironment)
+
+
+@pytest.fixture()
+def launcher(app):
+    return Launcher(Config(), app)
+
+
+@pytest.fixture()
+def process(launcher):
+    launcher._spawn_process({"_project": "p1", "_spider": "s1", "_job": "j1"}, 0)  # noqa: SLF001
+    return launcher.processes[0]
 
 
 @pytest.mark.parametrize(
@@ -17,3 +45,113 @@ from scrapyd.launcher import get_crawl_args
 )
 def test_get_crawl_args(message, expected):
     assert get_crawl_args(message) == expected
+
+
+def test_start_service(launcher):
+    with capturedLogs() as captured:
+        launcher.startService()
+
+    assert len(captured) == 1
+    assert captured[0]["log_level"] == LogLevel.info
+    assert re.search(
+        f"\\[Launcher\\] Scrapyd {__version__} started: max_proc=\\d+, runner='scrapyd.runner'", message(captured)
+    )
+
+
+def test_start_service_max_proc(app):
+    config = Config()
+    config.cp.set(Config.SECTION, "max_proc", "8")
+    launcher = Launcher(config, app)
+
+    with capturedLogs() as captured:
+        launcher.startService()
+
+    assert len(captured) == 1
+    assert captured[0]["log_level"] == LogLevel.info
+    assert re.search(
+        f"\\[Launcher\\] Scrapyd {__version__} started: max_proc=8, runner='scrapyd.runner'", message(captured)
+    )
+
+
+def test_out_received(process):
+    with capturedLogs() as captured:
+        process.outReceived(b"out\n")
+
+    assert len(captured) == 1
+    assert captured[0]["log_level"] == LogLevel.info
+    assert message(captured) == f"[Launcher,{process.pid}/stdout] out"
+
+
+def test_err_received(process):
+    with capturedLogs() as captured:
+        process.errReceived(b"err\n")
+
+    assert len(captured) == 1
+    assert captured[0]["log_level"] == LogLevel.error
+    assert message(captured) == f"[Launcher,{process.pid}/stderr] err"
+
+
+def test_connection_made(environ, process):
+    pid = process.pid
+    with capturedLogs() as captured:
+        process.connectionMade()
+
+    assert len(captured) == 1
+    assert captured[0]["log_level"] == LogLevel.info
+    if environ.items_dir:
+        assert re.match(
+            f"\\[scrapyd\\.launcher#info\\] Process started: project='p1' spider='s1' job='j1' pid={pid} "
+            "args=\\['\\S+', '-m', 'scrapyd\\.runner', 'crawl', 's1', '-s', 'LOG_FILE=\\S+j1\\.log', '-s', "
+            """'FEEDS={"file://\\S+j1\\.jl": {"format": "jsonlines"}}', '-a', '_job=j1'\\]""",
+            message(captured),
+        )
+    else:
+        assert re.match(
+            f"\\[scrapyd\\.launcher#info\\] Process started: project='p1' spider='s1' job='j1' pid={pid} "
+            "args=\\['\\S+', '-m', 'scrapyd\\.runner', 'crawl', 's1', '-s', 'LOG_FILE=\\S+j1\\.log', '-a', '_job=j1'\\]",
+            message(captured),
+        )
+
+
+def test_process_ended_done(environ, process):
+    pid = process.pid
+    with capturedLogs() as captured:
+        process.processEnded(failure.Failure(error.ProcessDone(0)))
+
+    assert len(captured) == 1
+    assert captured[0]["log_level"] == LogLevel.info
+    if environ.items_dir:
+        assert re.match(
+            f"\\[scrapyd\\.launcher#info\\] Process finished: project='p1' spider='s1' job='j1' pid={pid} "
+            "args=\\['\\S+', '-m', 'scrapyd\\.runner', 'crawl', 's1', '-s', 'LOG_FILE=\\S+j1\\.log', '-s', "
+            """'FEEDS={"file://\\S+j1\\.jl": {"format": "jsonlines"}}', '-a', '_job=j1'\\]""",
+            message(captured),
+        )
+    else:
+        assert re.match(
+            f"\\[scrapyd\\.launcher#info\\] Process finished: project='p1' spider='s1' job='j1' pid={pid} "
+            "args=\\['\\S+', '-m', 'scrapyd\\.runner', 'crawl', 's1', '-s', 'LOG_FILE=\\S+j1\\.log', '-a', '_job=j1'\\]",
+            message(captured),
+        )
+
+
+def test_process_ended_terminated(environ, process):
+    pid = process.pid
+    with capturedLogs() as captured:
+        process.processEnded(failure.Failure(error.ProcessTerminated(1)))
+
+    assert len(captured) == 1
+    assert captured[0]["log_level"] == LogLevel.error
+    if environ.items_dir:
+        assert re.match(
+            f"\\[scrapyd\\.launcher#error\\] Process died: exitstatus=1 project='p1' spider='s1' job='j1' pid={pid} "
+            "args=\\['\\S+', '-m', 'scrapyd\\.runner', 'crawl', 's1', '-s', 'LOG_FILE=\\S+j1\\.log', '-s', "
+            """'FEEDS={"file://\\S+j1\\.jl": {"format": "jsonlines"}}', '-a', '_job=j1'\\]""",
+            message(captured),
+        )
+    else:
+        assert re.match(
+            f"\\[scrapyd\\.launcher#error\\] Process died: exitstatus=1 project='p1' spider='s1' job='j1' pid={pid} "
+            "args=\\['\\S+', '-m', 'scrapyd\\.runner', 'crawl', 's1', '-s', 'LOG_FILE=\\S+', '-a', '_job=j1'\\]",
+            message(captured),
+        )